+++ examples/complete/versions.tf
@@ -1,0 +1,18 @@
+terraform {
+  required_version = ">= 1.0"
+
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 6.0"
+    }
+    github = {
+      source  = "integrations/github"
+      version = "~> 6.0"
+    }
+    gitlab = {
+      source  = "gitlabhq/gitlab"
+      version = "~> 3.0"
+    }
+  }
+}